fix(e2e): Change to setInterval to run outside of angular zone
The call to setInterval inside the angular zone makes
it appear that angular is constantly busy, protractor
implicitly calls waitForAngular as part of interacting
with the website, which hangs the test. This fix tells
it to run the setInterval outside of angular's zone,
allowing E2E tests to run.
Fix based on
https://github.com/HackedByChinese/ng2-idle/commit/d8599d7